docs/source/Analytics/OpenSearch-Service-Root/02-OpenSearch-Serverless-Version/deploy_cdk_requirements.txt
aws-cdk-lib==2.81.0
constructs==10.2.70
